+++ TicTacToe/Cross.cpp
@@ -1,9 +1,10 @@
 #include "Engine/BaseTypes.h"
-#include "glm/glm.hpp"
+#include "glm.hpp"
 #include "Cross.h"
 
 #include <vector>
-#include "gl/glew.h"
+#include "gtc/matrix_transform.hpp"
+#include "glew.h"
 
 #include "Engine/GLSLProgram.h"
 #include "Engine/Camera2D.h"
@@ -58,7 +59,7 @@ void Cross::buildModel()
 		0, 1, 2, 0, 2, 3,
 		4, 5, 6, 4, 6, 7
 	});
-	iboSize = indices.size();
+	iboSize = (uint)indices.size();
 
 	//upload to GPU
 	glGenBuffers(1, &vboId);
